requirements.txt
torch>=1.10.0
torchvision>=0.11.1
transformers>=4.18.0
pillow>=9.0.0
numpy>=1.22.0
pandas>=1.4.0
matplotlib>=3.5.0
scikit-learn>=1.0.2
tqdm>=4.62.3
tensorboard>=2.8.0
jsonlines>=3.0.0
wandb>=0.12.0
seaborn>=0.11.2
opencv-python>=4.5.5
albumentations>=1.1.0
timm>=0.5.4
jupyter>=1.0.0
ipykernel>=6.9.0 